backend/requirements.txt
asgiref==3.4.1
Django==3.2.6
gunicorn==20.1.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
whitenoise==5.2.0

django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
Markdown==3.3.4
PyJWT==2.1.0